# docs/api/hair-tryon-service-openapi.yaml
openapi: 3.0.3
info:
  title: GrowUp Hair Try-On Service API
  description: AI-powered hair style try-on service with video processing and real-time streaming for the GrowUp mobile application
  version: 1.0.0
  contact:
    name: GrowUp Development Team
    email: dev@growup.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3004/api/hair
    description: Development server
paths:
...
  /history/{user_id}:
    get:
      summary: Get user processing history
      description: Retrieve user's hair try-on processing history
      tags:
        - History
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of results to skip
        - name: type_filter
          in: query
          schema:
            type: string
            enum: ["video", "realtime"]
          description: Filter by processing type
      responses:
        '200':
          description: History retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to retrieve history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /result/{result_id}:
    delete:
      summary: Delete processing result
      description: Delete a hair try-on result and associated files
      tags:
        - Results
      parameters:
        - name: result_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Result ID to delete
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: User ID for authorization
      responses:
        '200':
          description: Result deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Result not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to delete result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    HistoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "result_abc123de-e89b-12d3-a456-426614174003"
        type:
          type: string
          enum: ["video", "realtime"]
          example: "video"
        status:
          type: string
          enum: ["pending", "processing", "completed", "failed"]
          example: "completed"
        result_media_url:
          type: string
          nullable: true
          example: "/uploads/result_abc123de_result.mp4"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        processing_time:
          type: number
          nullable: true
          example: 45.2
    HistoryResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        results:
          type: array
          items:
            $ref: '#/components/schemas/HistoryItem'
        pagination:
          type: object
          properties:
            limit:
              type: integer
              example: 20
            offset:
              type: integer
              example: 0
            total:
              type: integer
              example: 45
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "PROCESSING_FAILED"
        message:
          type: string
          example: "Video processing failed"
        details:
          type: string
          nullable: true
          example: "AI model inference error: insufficient GPU memory"
...